requirements.txt
apns2==0.7.2
appdirs==1.4.4
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
DateTime==4.3
distlib==0.3.2
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-push-notifications==2.0.0
django-truncate==0.1
django-upgrade==1.4.0
django-webpush==0.3.3
djongo==1.3.6
dnspython==2.1.0
filelock==3.0.12
geographiclib==1.52
geopy==2.1.0
gunicorn==20.1.0
h2==2.6.2
haversine==2.3.1
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
jsonfield==3.1.0
mongoengine==0.23.1
psycopg2==2.9.1
py-vapid==1.8.2
pycparser==2.20
PyJWT==1.7.1
pymongo==3.12.3
python-decouple==3.5
pytz==2021.1
pywebpush==1.9.4
requests==2.25.1
six==1.16.0
sqlparse==0.2.4
tokenize-rt==4.2.1
tzdata==2021.5
urllib3==1.26.6
virtualenv==20.4.7
whitenoise==5.2.0
xlwt==1.3.0
zope.interface==5.4.0
